refactor: SectionLabel to shared
