+++ src/app/components/index.ts
@@ -8,6 +8,5 @@ export { OptionItem, type OptionItemProps } from './option-item';
 export { Places } from './places';
 export { PrincipleCard, type PrincipleCardProps } from './principle-card';
 export { SchoolMenu } from './school-menu';
-export { SectionLabel } from './section-label';
 export { SocialMedia, type SocialMediaProps } from './social-media';
 export { TextWithLink } from './text-with-link';
