+++ src/shared/ui/section-label/index.ts
