+++ src/shared/ui/section-label/section-label.tsx
